requirements.txt
aiohttp==3.13.2
httpx==0.28.1
playwright==1.56.0
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
rich==14.2.0
